+++ GenericModConfigMenu/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* The generic config UI is now translatable. (Option labels added by other mods aren't translated yet.)
+
 ## 1.4.2
 Released 11 September 2021 for SMAPI 3.12.5 or later. Updated by Pathoschild and spacechase0.